fix: when activity dropdown is expanded, don't restrict number of lines for goal indicators
