+++ lib/routes/chat/activity_sessions/activity_sessions_start_view.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
 import 'package:fluffychat/config/themes.dart';
-import 'package:fluffychat/features/activity_sessions/activity_plan_model.dart';
 import 'package:fluffychat/features/navigation/panel_token.dart';
 import 'package:fluffychat/features/navigation/room_id_url.dart';
 import 'package:fluffychat/features/navigation/route_facts.dart';
@@ -12,13 +11,13 @@ import 'package:fluffychat/features/navigation/workspace_nav.dart';
 import 'package:fluffychat/features/navigation/workspace_query.dart';
 import 'package:fluffychat/l10n/l10n.dart';
 import 'package:fluffychat/pangea/common/widgets/error_indicator.dart';
+import 'package:fluffychat/routes/chat/activity_sessions/activity_goals_dropdown.dart';
 import 'package:fluffychat/routes/chat/activity_sessions/activity_participant_list.dart';
 import 'package:fluffychat/routes/chat/activity_sessions/activity_session_bottom_content.dart';
 import 'package:fluffychat/routes/chat/activity_sessions/activity_session_button_widget.dart';
 import 'package:fluffychat/routes/chat/activity_sessions/activity_session_start_page.dart';
 import 'package:fluffychat/routes/chat/activity_sessions/activity_session_state_controller.dart';
 import 'package:fluffychat/routes/chat/activity_sessions/activity_vocab_widget.dart';
-import 'package:fluffychat/routes/chat/choreographer/activity_orchestrator/goal_status_widget.dart';
 import 'package:fluffychat/utils/stream_extension.dart';
 import 'package:fluffychat/widgets/matrix.dart';
 import 'package:fluffychat/widgets/url_image_widget.dart';
@@ -280,7 +279,7 @@ class ActivitySessionStartView extends StatelessWidget {
                                     top: 0,
                                     left: 0,
                                     right: 0,
-                                    child: _ActivityGoalsDropdown(
+                                    child: ActivityGoalsDropdown(
                                       goals:
                                           sessionController.selectedRoleGoals,
                                       completedGoalIds: sessionController
@@ -351,161 +350,6 @@ class ActivitySessionStartView extends StatelessWidget {
           ),
         );
       },
-    );
-  }
-}
-
-class _ActivityGoalsDropdown extends StatefulWidget {
-  final List<ActivityRoleGoal>? goals;
-  final Set<String> completedGoalIds;
-  final bool startCollapsed;
-
-  const _ActivityGoalsDropdown({
-    required this.goals,
-    required this.completedGoalIds,
-    this.startCollapsed = false,
-  });
-
-  @override
-  State<_ActivityGoalsDropdown> createState() => _ActivityGoalsDropdownState();
-}
-
-class _ActivityGoalsDropdownState extends State<_ActivityGoalsDropdown> {
-  bool _visible = false;
-  bool _innerExpanded = true;
-  List<ActivityRoleGoal>? _displayGoals;
-
-  @override
-  void initState() {
-    super.initState();
-    _displayGoals = widget.goals;
-    _visible = widget.goals != null && widget.goals!.isNotEmpty;
-    _innerExpanded = !widget.startCollapsed;
-  }
-
-  @override
-  void didUpdateWidget(covariant _ActivityGoalsDropdown old) {
-    super.didUpdateWidget(old);
-    if (old.goals != widget.goals) {
-      final hasGoals = widget.goals != null && widget.goals!.isNotEmpty;
-      if (hasGoals) {
-        setState(() {
-          _displayGoals = widget.goals;
-          _visible = true;
-          _innerExpanded = true;
-        });
-      } else {
-        setState(() => _visible = false);
-        Future.delayed(FluffyThemes.animationDuration, () {
-          if (mounted) setState(() => _displayGoals = null);
-        });
-      }
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final goals = _displayGoals ?? [];
-    final theme = Theme.of(context);
-    final firstGoal = goals.isNotEmpty ? goals.first : null;
-    final remainingGoals = goals.length > 1
-        ? goals.skip(1).toList()
-        : <ActivityRoleGoal>[];
-
-    return ClipRect(
-      child: AnimatedAlign(
-        duration: FluffyThemes.animationDuration,
-        curve: Curves.easeInOut,
-        heightFactor: _visible ? 1.0 : 0.0,
-        alignment: Alignment.topCenter,
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            if (firstGoal != null)
-              InkWell(
-                onTap: remainingGoals.isNotEmpty
-                    ? () => setState(() => _innerExpanded = !_innerExpanded)
-                    : null,
-                child: Container(
-                  height: 55.0,
-                  padding: const EdgeInsets.symmetric(horizontal: 12.0),
-                  decoration: BoxDecoration(
-                    color: theme.colorScheme.surface,
-                    border: Border(top: BorderSide(color: theme.dividerColor)),
-                  ),
-                  child: Row(
-                    children: [
-                      Expanded(
-                        child: GoalStatusWidget(
-                          goal: firstGoal,
-                          complete: widget.completedGoalIds.contains(
-                            firstGoal.id,
-                          ),
-                        ),
-                      ),
-                      if (remainingGoals.isNotEmpty)
-                        Icon(
-                          _innerExpanded
-                              ? Icons.expand_less
-                              : Icons.expand_more,
-                        ),
-                    ],
-                  ),
-                ),
-              ),
-            if (remainingGoals.isNotEmpty)
-              ClipRect(
-                child: AnimatedAlign(
-                  duration: FluffyThemes.animationDuration,
-                  curve: Curves.easeInOut,
-                  heightFactor: _innerExpanded ? 1.0 : 0.0,
-                  alignment: Alignment.topCenter,
-                  child: GestureDetector(
-                    onPanUpdate: (d) {
-                      if (d.delta.dy < -2) {
-                        setState(() => _innerExpanded = false);
-                      }
-                    },
-                    child: LayoutBuilder(
-                      builder: (context, constraints) {
-                        final maxHeight =
-                            MediaQuery.of(context).size.height * 0.7;
-                        return ConstrainedBox(
-                          constraints: BoxConstraints(maxHeight: maxHeight),
-                          child: SingleChildScrollView(
-                            child: Container(
-                              width: double.infinity,
-                              color: theme.colorScheme.surface,
-                              padding: const EdgeInsets.fromLTRB(
-                                12.0,
-                                12.0,
-                                12.0,
-                                24.0,
-                              ),
-                              child: Column(
-                                spacing: 16.0,
-                                mainAxisSize: MainAxisSize.min,
-                                children: remainingGoals
-                                    .map(
-                                      (g) => GoalStatusWidget(
-                                        goal: g,
-                                        complete: widget.completedGoalIds
-                                            .contains(g.id),
-                                      ),
-                                    )
-                                    .toList(),
-                              ),
-                            ),
-                          ),
-                        );
-                      },
-                    ),
-                  ),
-                ),
-              ),
-          ],
-        ),
-      ),
     );
   }
 }
